+++ smartsheet/models/report.py
@@ -19,7 +19,7 @@ from __future__ import absolute_import
 
 from ..types import EnumeratedValue, TypedList, TypedObject, json
 from ..util import deserialize, serialize
-from .enums import ResourceType
+from .enums import ChildrenResourceType
 from .report_column import ReportColumn
 from .report_row import ReportRow
 from .scope import Scope
@@ -38,7 +38,7 @@ class Report(Sheet):
             self._base = base_obj
 
         self._columns = TypedList(ReportColumn)
-        self._resource_type = EnumeratedValue(ResourceType)
+        self._resource_type = EnumeratedValue(ChildrenResourceType)
         self._rows = TypedList(ReportRow)
         self._scope = TypedObject(Scope)
         self._source_sheets = TypedList(Sheet)